Fix type

+++ src/xtd.forms/src/xtd/forms/loading_indicator_animation_circle_blinks.h
@@ -6,7 +6,7 @@ namespace xtd::forms {
   class loading_indicator_animation_circle_blinks : public loading_indicator_animation {
   public:
     void on_paint(xtd::drawing::graphics& graphics, const xtd::drawing::rectangle& clip_rectangle, const xtd::drawing::color& color, bool enabled) override {
-      graphics.translate_transform(clip_rectangle.center().x(), clip_rectangle.center().y());
+      graphics.translate_transform(static_cast<float>(clip_rectangle.center().x()), static_cast<float>(clip_rectangle.center().y()));
       auto angle = -360.f / intervals_;
       graphics.rotate_transform(frame_ * angle);
       auto height = static_cast<float>(xtd::math::min(clip_rectangle.width(), clip_rectangle.height())) / radius_factor_;
